refactor thank you body
