+++ lib/features/payment/presentation/widgets/thank_you_card.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/cupertino.dart';
+
+class ThankYouCard extends StatelessWidget {
+  const ThankYouCard({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: ShapeDecoration(
+        color: const Color(0xffD9D9D9),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(20),
+        ),
+      ),
+    );
+  }
+}
