+++ lib/features/payment/presentation/widgets/total_price_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import '../../../../core/utlis/styles.dart';
+import '../../../../core/utils/styles.dart';
 
 class TotalPrice extends StatelessWidget {
   const TotalPrice({super.key, required this.title, required this.value});
